fix: Share media card metadata helpers (#711)
