+++ frontend/src/features/video-viewer/components/AutoPlay.jsx
@@ -9,41 +9,11 @@ import { HorizontalMovieItem } from '../../shared/components/MovieItem/MovieItem
 import { Switch } from '../../shared/components/Switch/Switch';
 import { Text } from '../../shared/components/Text/Text';
 import { getMediaDurationLabel } from '../../home/utils/mediaList';
+import { buildMetadata, getAuthorLink, getAuthorName } from '../utils/mediaCardMetadata';
 
 function readAutoPlayMedia() {
 	const md = MediaPageStore.get('media-data');
 	return md?.related_media?.length ? md.related_media[0] : null;
-}
-
-function getAuthorName(item) {
-	return item.author_name || item.user || '';
-}
-
-function getAuthorLink(item) {
-	return item.author_profile || '';
-}
-
-function getCountryNames(item) {
-	if (Array.isArray(item.media_country_info)) {
-		const countries = item.media_country_info.map((country) => country?.title).filter(Boolean);
-		return countries.length ? countries : [item.media_country].filter(Boolean);
-	}
-
-	if (item.media_country_info?.title) {
-		return [item.media_country_info.title];
-	}
-
-	return [item.media_country].filter(Boolean);
-}
-
-function buildMetadata(item, hideViews) {
-	const metadata = [...getCountryNames(item)];
-
-	if (!hideViews && item.views != null) {
-		metadata.push(`${Number(item.views).toLocaleString()} views`);
-	}
-
-	return metadata;
 }
 
 export function AutoPlay() {
